Add files via upload
This is the revised version of the meeting manager code, class diagram and still in progress method descriptions.

What's different from previous version:
1. The MeetingManager class holds the methods that do all the processing
2. Employee, Meeting, MeetingSuggestion, MeetingSuggestionMade and Notification classes are solely used to hold information and
only consist of setter and getter methods.
3. All the methods in the MeetingManager take in a reference of an already constructed instance of a Meeting for whatever purpose; instead of taking in details separately and instantiating the Meeting inside the method. This means that when we have a menu, we create the instance of a Meeting in the menu option and pass that into the method if necessary.

I'm not sure point 3 makes much sense but you'll get the idea when you look at the code already there.

Ultimately, right now you should only be making changes to the MeetingManager class as the other are pretty much in their final state.

As you'll see a lot of the methods don't actually have code in them but have very thorough descriptions of their purpose and functionality, with parameters already in place. so if you should probably start with those. I'll get the descriptions for the other ones done soon as well.